+++ test/phoenix_flags_test.exs
@@ -349,12 +349,4 @@ defmodule PhoenixFlags.ServerTest do
       assert TestConfig.benefits_enabled?()
     end
   end
-
-  defp errors_on(changeset) do
-    Ecto.Changeset.traverse_errors(changeset, fn {message, opts} ->
-      Regex.replace(~r"%{(\w+)}", message, fn _, key ->
-        opts |> Keyword.get(String.to_existing_atom(key), key) |> to_string()
-      end)
-    end)
-  end
 end